+++ generators/csharp/codegen/src/custom-config/BaseCsharpCustomConfigSchema.ts
@@ -2,7 +2,8 @@ import { z } from "zod";
 
 export const BaseCsharpCustomConfigSchema = z.object({
     namespace: z.string().optional(),
-    "simplify-object-dictionaries": z.boolean().optional()
+    "simplify-object-dictionaries": z.boolean().optional(),
+    "read-only-memory-types": z.optional(z.array(z.string()))
 });
 
 export type BaseCsharpCustomConfigSchema = z.infer<typeof BaseCsharpCustomConfigSchema>;