+++ seed/csharp-sdk/audiences/src/SeedAudiences/FolderA/Service/Types/Response.cs
@@ -1,4 +1,5 @@
 using System.Text.Json.Serialization;
+using SeedAudiences.FolderB;
 
 #nullable enable
 
@@ -7,5 +8,5 @@ namespace SeedAudiences.FolderA;
 public record Response
 {
     [JsonPropertyName("foo")]
-    public FolderB.Foo? Foo { get; set; }
+    public Foo? Foo { get; set; }
 }