+++ seed/csharp-sdk/cross-package-type-names/src/SeedCrossPackageTypeNames/Commons/CommonsClient.cs
@@ -1,0 +1,15 @@
+using SeedCrossPackageTypeNames.Core;
+
+#nullable enable
+
+namespace SeedCrossPackageTypeNames;
+
+public partial class CommonsClient
+{
+    private RawClient _client;
+
+    internal CommonsClient(RawClient client)
+    {
+        _client = client;
+    }
+}