+++ seed/csharp-sdk/cross-package-type-names/src/SeedCrossPackageTypeNames/FolderB/Common/Types/Foo.cs
@@ -1,0 +1,11 @@
+using System.Text.Json.Serialization;
+
+#nullable enable
+
+namespace SeedCrossPackageTypeNames.FolderB;
+
+public record Foo
+{
+    [JsonPropertyName("foo")]
+    public FolderC.Foo? Foo_ { get; set; }
+}