+++ seed/csharp-sdk/cross-package-type-names/src/SeedCrossPackageTypeNames/Foo/Types/ImportingType.cs
@@ -1,0 +1,11 @@
+using System.Text.Json.Serialization;
+
+#nullable enable
+
+namespace SeedCrossPackageTypeNames;
+
+public record ImportingType
+{
+    [JsonPropertyName("imported")]
+    public required string Imported { get; set; }
+}