+++ seed/ruby-sdk/audiences/lib/environment.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module SeedAudiencesClient
+  class Environment
+    ENVIRONMENT_A = "https://api.example.a.com"
+    ENVIRONMENT_B = "https://api.example.b.com"
+  end
+end