+++ seed/ts-sdk/cross-package-type-names/src/index.ts
@@ -1,0 +1,3 @@
+export * as SeedCrossPackageTypeNames from "./api";
+export { SeedCrossPackageTypeNamesClient } from "./Client";
+export { SeedCrossPackageTypeNamesError, SeedCrossPackageTypeNamesTimeoutError } from "./errors";